Experiment with promise.fold
This impl threads the extra arg through the handler chain. This is
good in that it completely avoids any extra closures.  It adds an
extra threading arg to handlers, and an check to
FulfilledHandler.prototype.when, but that seems to have no observable
effect on performance.  On the upside, it allows all() to avoid
capturing the results array, instead threading it as the extra arg