Experiment with using an object to thread continuation params through handler.when
In some microbenchmarks, this has almost no noticeable effect on
performance.  In a few other, more real-world perf tests, though, it
seems to make up to a 10% diff in perf, but only a very small diff in
mem usage.  That's not too bad.  If we could find a way to bring that
down a bit, this could be worthwhile.  I didn't really spend any time
trying to optimize this version, so it might be possible.

I think I was a bit smarter about implementing it this time than in
previous attempts.  The object literals used for continuations are all
monomorphic this time, so the VM can optimize around their shape more
effectively.

cc @jenius